Hide playlist from ranked play rooms
This also implements the concept of "effects". Every card has an
associated effect that is activated when the card itself is activated.

In order to hide playlist items from the rooms, I'm inserting a dummy
playlist item when the room is created, which is required for clients to
behave correctly (CurrentPlaylistItem cannot be null).

When a card is played, I'm either updating that dummy playlist item (if
it's not expired), or adding a new playlist item.